Rework partitioning to drop LVM
This is part of https://github.com/coreos/fedora-coreos-tracker/issues/18

For now, this just drops LVM to make it easier to use Ignition
to both build images, and help enable ignition-disks.

Note that I tried to use a separate `/var` but this currently
does not work with our Ignition, which would need to learn
how to mount `/var` in the initramfs.

We add growpart logic adapted from
https://github.com/projectatomic/container-storage-setup/pull/275/commits/d4994e68408f7a715d7741b25fcf9c8332b09414

(Probably at some point should teach growpart how to grow based
 on mount point paths...)