image: Use reflinks for / by default
This is transparently used by overlayfs for copyups, rpm-ostree
can use it for the rpmdb, and ostree uses it for `/etc`.

It's The Future.  (Also probably the default for `mkfs.xfs` at
some point I imagine...if that happens we'll need to explicitly
turn it off for `/boot` or perhaps suck it up and use `ext4`
or something)

https://lwn.net/Articles/659677/
https://lwn.net/Articles/747633/