base: Use /var/home for home directories by default
This came up in `#silverblue` discussion; let's do it here
first under the general concept that most changes to the "core"
underlying both Silverblue and FCOS should be made in FCOS first.

Tested locally with a build+run.  It doesn't affect the `core`
user today, because anaconda+libuser aren't reading the defaults
from the target OS.