Add coreos-useradd-core.service to create `core` user
I'd like to change coreos-assembler to have `/var` completely
empty in the disk image, and hence all content is created consistently
via systemd-tmpfiles at boot.

This is an important preparatory step for moving away from Anaconda.

For backwards compatibility (for now) let's auto-create the `core`
user still, but via a systemd unit on boot.  The right thing
I think is to move this logic into Ignition, but this works
for now.