Add image.yaml
Let's just support configuring the base disk size for now so that we
say we *do* have some configuration.  We need to abstract this for
multiple reasons (we want to remove Anaconda, and even without removing
Anaconda we want to do dual bios/UEFI and also potentially arch-specific
code).

Most people will probably be OK for now configuring the base ostree and the
outer image size.

This commit doesn't yet remove image.ks in order to support version skew,
but we intend to shortly drop it.