overlay: Seed /root with default bash profile
This has a long history in Atomic Host:
https://bugzilla.redhat.com/show_bug.cgi?id=1193590

The TL;DR is that since OSTree commits ship with empty `/var`, and
`/root` is really `/var/roothome`, then the default `.bash*` files for
the root don't actually exist. The RHBZ above was to try to teach bash
to fallback to the default files automatically, but it was deemed too
intrusive in the end and backed out.

In Atomic Host, we've worked around this by simply seeding the files
from the kickstart:
https://pagure.io/fedora-kickstarts/c/740a9b4969ca30da36458faf6809908f4da904c4?branch=master

This patch does something similar but at Ignition time. We do this
before the files stage so that for all purposes, this is part of the
distro (e.g. an Ignition config can append to `/root/.bashrc` for
example, or can add a file to `/etc/profile.d/` and expect it to be
sourced).

Concretely, this fixes non-login shells not sourcing `/etc/bashrc` and
`/etc/profile.d/` and thus having a prompt like `bash-5.0# `, no `ls`
colouring, non-default umask, etc...