repos: move to coreos-pool repo
This is an intermediate step towards our longer term goal of consuming
everything for our production refs from a distRepo generated from the
coreos-pool koji tag. Right now we'll just consume from the coreos-pool
as well as the traditional fedora repos.

This replaces the fedora-coreos-continuous repo as that will be used for
CI streams that rebuild quickly from upstream sources. We were currently
using it to bypass bodhi but we can use coreos-pool for that too.