fedora.repo: add modular repos
This is prep for having `testing-devel` be the source of truth for all
the files other than the lockfile and manifest (see [1]). We just have
one `fedora.repo` file which contains all the repos we care about, but
only actually enable the appropriate ones for each stream independently.

In this case, once packages are tagged, `testing-devel` will just use
the `coreos-pool` repo.

[1] https://github.com/coreos/fedora-coreos-tracker/issues/204