base: Tweak versioning to conform to new spec
In https://github.com/coreos/fedora-coreos-tracker/issues/81, we agreed
on a versioning scheme for releases. This patch changes
`automatic-version-prefix` to respect this.

Note that it's not a perfect fit, since e.g. the first build doesn't
have a `.0` prefix.

In actuality, versioning in the pipeline will be driven by an external
process, and thus the scheme embedded here will have no effect (unless
we decide to use rpm-ostree itself as part of the external process...
though it'll need some tweaking first).

So really, this is more for the benefit of local developers so that
versioning looks similar to release builds.