base: add .dev marker by default
We want to be able to distinguish between release and non-release
builds. To do this, we embed a `.dev` marker in the version string so
that the default developer workflow will use this. As mentioned in the
previous commit message, this will be overridden for release streams.

See https://github.com/coreos/fedora-coreos-tracker/issues/212 for more
information about this.