Add coreos-update-ca-trust.service
This blends well with Ignition, allowing people to just drop
in certs without having to worry about injecting their own
systemd service to update them.

We also want this for RHCOS.
See: https://github.com/openshift/machine-config-operator/issues/528
For FCOS, in theory this could be a "run at most once" service.
But for RHCOS, since we support dynamic changes, we need it to
run on each boot too.  The cost of doing so is very small.