overlay/05core: Remove /etc/profile.d/path.sh
Trying to rebase RHCOS on this config, I hit upon a conflict
because long ago (before `overlay/` existed) we stuck this
in the `redhat-release-coreos` package.

One thing I wanted to verify is; is this really necessary for
FCOS today?  And the answer is: no.  The original rationale
from https://github.com/openshift/os/issues/191 mentioned the
kola SELinux tests, but `cosa kola` passes just fine with this
as is.