README.md: add instructions on adding new packages
Now that we've completely switched over to lockfiles, we need to
describe the procedure for adding new packages to the OS.

Also add back the Bodhi repos to make the task easier. Otherwise, we'd
get a bootstrapping problem: we can't add packages not tagged in
coreos-pool, but the tagger won't tag a package that's not added.

But also, we really want those repos there by default so that the local
dev case of pulling in all new packages can just be
`cosa fetch --update-lockfile`.