overlay: remove 40coreos-network dracut module
Remove the 40coreos-network, which was intended to bring down
initramfs networking. Currently this module wasn't enabled anyways,
and will become obsolete once we move to NetworkManager based
initramfs networking.

For context, this is currently used in RHCOS to teardown networking
before the switch root, to account for static use cases. This
effectively allowed users to dhcp in the initramfs to fetch ignition
configs, and then have ignition set up static networking on the real
root. Without this teardown it will instead just inherit initramfs
networking.

Signed-off-by: Yu Qi Zhang <jerzhang@redhat.com>