overlay: add Azure udev rules
The downstream RHBZ#1747575[0] requests the addition of these udev
rules to allow nodes in Azure to handle the dynamic provisioning of
storage devices.  The WALinuxAgent repo[1] provides the two udev rules
required for this functionality, which haven't changed in a year or more.
This commit is just a copy of the rules from the repo.

[0] https://bugzilla.redhat.com/show_bug.cgi?id=1747575
[1] https://github.com/Azure/WALinuxAgent/tree/master/config/