overlay: add GCP udev rules
The downstream BZ#1748638[0] requests the addition of udev rules for
nodes in GCP to support the dynamic provisioning of storage.  The
rules live in the `GoogleCloudPlatform/compute-image-packages` repo[1]
and were most recently updated in Dec 2018.

This commit is mostly a copy of those rules with additional scoping to
only run in the GCP environment.

[0] https://bugzilla.redhat.com/show_bug.cgi?id=1748638
[1] https://github.com/GoogleCloudPlatform/compute-image-packages/tree/master/packages/google-compute-engine/src/lib/udev/rules.d