05core: automatically log in on console, only when booted from ISO
Automatic login is insecure and we don't want to encourage its use.
However, the live ISO image is an exception.  The ISO image is more
useful for casual experimentation and for running coreos-installer if
a shell is readily available.  Conversely, passing an Ignition config
to the ISO requires manually modifying the kernel command line, so the
ISO is unlikely to be run in production.

Automatically log in on the graphical console when both a) booted
from a live image, and b) the kernel argument `iso` is supplied.
Modify live ISO bootloader configs to pass the argument.