manifest: Explicitly pull in kernel-core, not kernel-debug-core
We were just adding `kernel` which is a metapackage.  Recently
it looks like the kernel package grew a separate `kernel-debug-core`
(I'm guessing to make it easy to switch between the two)

We need to explicitly pull in `kernel-core` and `kernel-modules`
to ensure the depsolver doesn't pull in the debug kernel.