fedora-coreos-base:  add cifs-utils
In the downstream BZ#1752395[0], it is reported that users are unable
to mount Azure files in OCP 4.2.  This was a feature of OCP 3.11[1] and
we wish to maintain feature parity in OCP 4.2 on Azure as much as
possible.

This commit adds the `cifs-utils` package, in order to pull in the
`mount.cifs` binary to allow for the mounting of Azure files.

[0] https://bugzilla.redhat.com/show_bug.cgi?id=1752395
[1] https://docs.openshift.com/container-platform/3.11/architecture/additional_concepts/storage.html