Drop ostree-grub2
Since we switched to having the BLS be canonical and use the
`ostree.bootloader=none` config, we don't need the `15_ostree`
stuff.

Drop the `ostree-grub2` package and that cruft, finally achieving
what the intention was from the start in having that be a separate
subpackage that was intended as a short term hack.