05core: clear SSSD NSS cache on live boots
On live systems with persistent /var, Ignition needs to recreate users
(`core` in particular) on every boot.  But the SSSD NSS cache is in /var
and thinks the users already exist, which causes
`chroot /sysroot id $userid` to succeed, so Ignition skips useradd and
fails on usermod.  Delete the cache to avoid this.