manifest: enable container_use_cephfs
The downstream RHCOS product has a small amount of SELinux booleans
that are explicitly enabled in the manifest to allow for proper
operation in OpenShift.  The only delta that I found between FCOS and
RHCOS, was the `container_use_cephfs` boolean.  This allows containers
to mount Ceph volumes.  It was requested in RHBZ#1692369[0].

[0] https://bugzilla.redhat.com/show_bug.cgi?id=1692369