overlay/ignition-ostree: Initial support for rootfs replacement
This adds basic infrastructure units for "re-provisioning"
the root filesystem.  See:
https://github.com/coreos/fedora-coreos-tracker/issues/94

A unit first detects if the Ignition configuration has a filesystem
with the label `root` - if so, we save the rootfs into RAM, let
`ignition-disks.service` run, then restore it from RAM.

Earlier attempts copied the files into `tmpfs`; this instead
uses the `brd` kernel module which is a RAM-backed block
device so we can just `dd`.  This a faster and more reliable way
to save the rootfs.  However, `brd` doesn't support discards,
so we require at minimum $rootfs_size RAM (e.g. 3G) until reprovisioning
is complete.  In the future I might investigate trying a `tmpfs`
again as that better integrates with the page cache and doesn't
have the discard issue.  But it requires more work to save the rootfs.

Future work here will likely more the `restore` phase into
`rpm-ostree`.