dracut: Add ignition-ostree-mount-sysroot.service
This is inert by default because coreos-assembler adds the `root=`
kernel argument today.  But the intention is to drop those and
allow integration with Ignition rootfs replacement.

We only want to pass the `prjquota` flag for our default XFS filesystem,
so this can't be a static `[Mount]` unit type, it has to be determined
after `ignition-fetch.service`.