fedora-coreos-base: drop nm-initrd-generator
In Fedora 31, NetworkManager includes an initrd module which replaces
the legacy dracut networking one. The key bit is here:

https://github.com/dracutdevs/dracut/blob/1fcc70fe57eea0ea658aa2de5c0044683fe85cf1/modules.d/40network/module-setup.sh#L11.

It was initially in f30 as well, but was reverted due to issues:
https://src.fedoraproject.org/rpms/NetworkManager/c/7d3054a9e366f1ebdd9d7854bf11a3069119d6da?branch=f30

(One thing that confused me at first: note that this isn't a systemd
generator, it's run via the dracut cmdline hook.)

The generator does seem to have feature parity wrt karg support:
https://github.com/NetworkManager/NetworkManager/blob/ba64c162dc471740a405f9d59e7a132f0e290078/src/initrd/nmi-cmdline-reader.c

Though I couldn't get it to work in my tests; `rd.neednet=1` does not
result in a populated `resolv.conf`, which would of course cause
Ignition to fail.

Anyway, given that we tie pretty strongly into initrd networking for
e.g. Ignition and the installer, I think I'd rather we stick with the
status quo for now, and do the transition separately from the FCOS f31
rebase.