overlay: Add a README noting deprecation to /etc/sysconfig
This directory is a legacy of the Red Hat Linux days; it
was an attempt to basically declare `/etc` as broken and
start a new "/etc in your /etc".  It didn't succeed.

The only reason we carry it forward to this day is inertia
and legacy.  Other distributions (all of the Debian derivatives
for example) don't ship it.

Let's add this file here starting with FCOS; we can bring
it to the other Fedora distributions later.