overlay: Bind ostree-prepare-root.service to ignition-complete.target
This unit comes from ostree upstream, but forms part of our
Ignition sequence.  Make it required by ignition-complete.target
so that if it fails we get the same OnFailure= behavior.

Making this change as during development `ignition-ostree-mount-sysroot.service`
wasn't running which caused this unit to fail; the root cause would have been
a bit more obvious with this change.