overlay: Tweak coreos-boot-mount-generator to deal with older RHCOS
I was debugging a customer's test system that was installed as
OpenShift 4.2 and then upgraded, and noticed rpm-ostreed wasn't
starting on reboot, which was because `/boot` wasn't mounted,
which was because the `systemd-fstab-generator` and this code
had actually raced and generated a corrupted `/run/systemd/boot.mount`.

Tweak this code so it cleanly exits and defers to systemd-fstab-generator
for RHCOS-4.2 or older systems.