overlay: Consistently use is-live-image in generators
This should close a race I hit while trying out a live system
(Silverblue-based-on-FCOS).

The generators here run in parallel; so we can't test for
`/run/ostree-live` in `coreos-boot-mount-generator` while
writing it in `/live-generator.`

Probably we should fold these generators together, but for
now it's clearer to use the `is-live-image` script anyways.