overlay: Move coreos-growpart to the initramfs
We discussed this in a few forums before (ignition-dracut PRs
and in person); there's no reason not to grow the rootfs
directly as part of our Ignition process.

This closes down all race conditions around having processes
writing to the storage before we grow it (e.g. ignition-files,
journald).

It also means that if growfs fails then provisioning fails, which
is what we want.

(One downside of this is AFAIK dracut doesn't have a way to
 say "take all dependencies of this RPM" so we have to manually
 do that.  Really at this point we should stick all of `coreutils`
 in the initramfs)