Add coreos-keep-cgroup-v1.service
In f31, the default cgroup changed to v2. However, we've decided to stay
on v1 for the time being. Thus, we don't want older nodes upgrading to
f31 to be forced into v2.

Add a tiny service which just scans the BLS configs and injects the
`systemd.unified_cgroup_hierarchy` karg as needed.

For more information, see:
https://github.com/coreos/fedora-coreos-tracker/issues/292
https://github.com/coreos/fedora-coreos-streams/issues/26#issuecomment-558361931