coreos-growpart: filter `findmnt` output through sort | uniq
Otherwise, if the device is e.g. mounted onto itself on the same path,
`findmnt` will print it twice. (This is better than just `head -n1`
because we still want to error out if somehow two or more *different*
devices are mounted at the same path on top of each other.