ignition-ostree-growfs.service: run before ostree-prepare-root
As mentioned in the comment there:

    This shouldn't be strictly necessary, but it's cleaner to not have
    OSTree muck around with moving mounts while we're still resizing the
    filesystem.

I initially thought that ignition-ostree-growfs was racing with
ostree-prepare-root and hence only why sometimes we'd get two listings
for `/sysroot` from the `findmnt` entry. But `ostree-prepare-root` here
AFAIK only does pure mount moves on `/sysroot`, so that shouldn't affect
it. Still, it feels cleaner to resize the rootfs upfront before we start
adding and moving mounts under it.