overlay: Add an ignition-ostree-generator to fix install path
The current documentation for the FCOS PXE install path:
https://docs.fedoraproject.org/en-US/fedora-coreos/bare-metal/#_live_pxe
Uses the `coreos.inst` kernel arguments instead of providing
an Ignition config.  In this case, we fail because
the `mount-subsequent.service` kicks in, and will fail to find
a root device.  (Or worse, may mistakenly find an existing disk
and mount it)

The bug here is a now familiar one - the
`ignition-ostree-mount-subsequent-sysroot.service` already has
`ConditionFileExists=!/run/ostree-live` which is correct for the
Live PXE case, but it also has a hard `Requires`, and as we know
systemd still pulls in requirements for units that aren't enabled.

Fix this by using a generator to enable the unit rather than doing
so as part of ignition-subsequent.

Note if Ignition is provided, then `ignition.firstboot` should be
present which would mean this bug isn't triggered.