overlay/ignition-ostree-growfs: Don't run if root= karg
This occurs in the currently artificial scenario of the RHCOS
upgrade tests from previous 4.2 Anaconda-based builds.  This wouldn't
happen in practice because growfs is already keyed off only executing
on the Ignition boot, and there's no way we'd rebase this code and
release updated 4.2 bootimages.  But, we already have similar exclusions in
the mount code, so let's do the same for growfs.