fedora-coreos-base: use NetworkManager for networking in the initramfs
Moving to NetworkManager in the initrd should help us solve some
problems we've been having with Networking. For what we want to
do in Fedora CoreOS doing this right requires the network to be
torn down in the initrd and also possibly propagated forward.

We effectively move *to* NetworkManager in the initrd (the default
in Fedora 31+) by putting back in the nm-initrd-generator, which
we were previously removing.

Requires: https://github.com/coreos/ignition-dracut/pull/159
Fixes: https://github.com/coreos/fedora-coreos-tracker/issues/394