tests: add a test that can be run manually for net testing
This is how I've been doing network testing for the NM in
the initrd PR (https://github.com/coreos/fedora-coreos-config/pull/310).
One day I'd like to convert this into some more sophisticated
test that can be run in our CI but AFAIU right now that's not
possibly because our CI is limited to qemu unprivileged networking.

For now let's put this $somewhere so that we can pull it out
at least when we suspect problems.