manifests: add NetworkManager-tui
`nmtui` allows for a user to interactively set up networking. Using Fedora
CoreOS interactively is usually not what we encourage with Fedora CoreOS, relying
on automation instead. The one exception to this is the Live ISO environment
where our current documented workflow is for a user to boot the Live ISO and
run coreos-installer interactively (of course, there is an automation workflow
as well). In this flow it would be a useful for a user to be able to have an
interactive tool that helped them set up networking if they need static networking.