ci: temporarily disable upgrade testing
We don't have a previous `next-devel` build yet that the upgrade tests
can use as the "parent" build. As mentioned in [1], we could force it to
use the testing-devel stream. But actually I'd like to enhance CI more
generally so it covers both of these cases (from `testing-devel` to
`next-devel` and within `next-devel` as well). For now, let's just focus
on getting f32-based FCOS off the ground.

And as mentioned in the Jenkinsfile, this will be automatically reverted
by config-bot when it next syncs the file from the `testing-devel`
branch.

[1] https://github.com/coreos/coreos-ci-lib/pull/19