manifest: drop fedora-coreos-pool repo
So... one thing I just realized is that now that we have a bunch of f32
packages in the pool, we really can't have `bodhi-updates` picking up
packages from there unlocked.

So this is sort of forcing our hand to get
https://github.com/coreos/fedora-coreos-tracker/issues/293 fixed very
soon so that we can unblock package bumps again.

As a first step and to prevent `testing-devel` ending up with a bunch of
f32 packages, let's remove the pool from the list of repos here.