live: disable audit on Live ISO
This is the nuclear option to fix the issue [1] where we have
audit messages coming to the console of our machines. We
specifically need the messages to be disabled on the Live ISO
because we are trying to support a workflow where a user can
interactively do an install and use TUI interfaces for configuring
networking. The increased user experience provided by a TUI is
completely lost if we have audit messages scrolling on the console
every 10 seconds or so.

[1] https://github.com/coreos/fedora-coreos-tracker/issues/220