manifest: include audit rpm
This change will cause the auditd daemon to get included in
Fedora CoreOS and be enabled by default. This is a roundabout
way of fixing [1] where we have audit messages coming to the
primary console of our FCOS machines.

This need for this is increased right now because we are trying
to support a workflow where a user can interactively do an install
using the Live ISO and use TUI interfaces for configuring networking.
The increased user experience provided by a TUI is completely lost
if we have audit messages scrolling on the console every 10 seconds
or so.

[1] https://github.com/coreos/fedora-coreos-tracker/issues/220