lockfiles: remove for now
F32 is more or less frozen right now. And it'd be really nice if we had
*something* f32-based in time for the f32 release. So let's just drop
all the lockfiles for now so we can do an unlocked build and avoid all
the issues that are still being fixed.

Our overrides should all already be in the pool, and since they're still
the highest NEVRA, they still get picked.