WIP: Enable initramfs-etc.img as an appended initramfs
The idea here is to enable configuration needed during the initramfs
without requiring the re-generation of it.

Previous models including [1] were considered. As counter-idea, it was
suggested to add to Dracut the ability to source in a signed CPIO
archive or at least checksum it (e.g. boot-etc=<file>:sha256) and then
rely on signed kargs. The chief concerns is that we need to ensure that
the content is machine specific and not enable an easy boot-foot gun.

This solution used Grub as the enforcement mechanism for both
authenticity and authentication of the change and bypasses Dracut
all-together. The core advantage is that admin could enable secure boot,
set the Grub password (to prevent menu editing) and then Grub would only
set "initramfs-etc='...'" if the file checked out.

[1] https://github.com/dracutdevs/dracut/pull/792

Signed-off-by: Ben Howard <ben.howard@redhat.com>