manifests: Move grub2-removals include to fedora-coreos
Move the include statement for grub2-removals from bootable-rpm-ostree
which is pulled in by fedora-coreos-base to the fedora-coreos manifest.

The exclusion of files in grub2-removals is specific to Fedora
CoreOS but not other potentially derivative operating systems:
- Fedora Silverblue uses `/etc/grub.d/10_reset_boot_success`
and `/etc/grub.d/12_menu_auto_hide`
- Fedora IoT uses `/etc/grub.d/08_fallback_counting` and
`/etc/grub.d/10_reset_boot_success`

This is a step towards making fedora-coreos-base re-usable as base
manifest for these operating systems.