Fix manifest lockfile extension from yaml to json
The lockfile mistakenly had a `.yaml` extension instead of `.json`. It
still worked fine of course, because rpm-ostree supports both, and cosa
looks for both.

Though it's misleading because the base lockfile is generated by
rpm-ostree itself, which technically generates JSON and not YAML (even
though JSON is in fact valid YAML).

So let's rename it. (We also have some tools today which would still
need to be adapted for handling YAML base lockfiles if we transitioned.)