20platform-chrony: Always resynchronize after drift of one second
This change is necessary to address https://bugzilla.redhat.com/show_bug.cgi?id=1780165
bug which talks about the system clock resychronization issue after VMs are suspended for
some time (close to 20 mins). In order to resolve this issue, we need to reconfigure the
chronyd setting (makestep 1.0 -1) to cause VMs to resync the time instantaneously.