overlay: Don't enable Live ISO networking w/injected config
We want people to be able to take our ISO and inject
Ignition so they can fully automate installs - all that
should be needed is to e.g. attach the ISO via an IPMI[1] system,
or written to a USB stick etc.

Today though, we enable the default DHCP if Ignition is
provided via `coreos-installer iso embed`, which breaks
automating static IP installs.

We've debated a lot of approaches for this, and general
consensus is that the best is probably "conditional networking"
in Ignition.  However that's a much larger amount of work that
also touches every artifact type.  This approach is
small and targeted and only affects the Live ISO path.

[1] https://en.wikipedia.org/wiki/Intelligent_Platform_Management_Interface