README.md: add `koji untag` to releasever bump SOP
We encountered this this morning: FCOS f30 signed packages were garbage
collected (even though the Koji builds themselves were not since they're
tagged in the pool), and this caused the `distRepo` task to fail because
it expects the signed packages to be there.

Ideally, we'd have proper GC for pool packages in place, but for now
let's add a step there to untag packages from the `N-2` release. This
interestingly does put a natural upper bound on how long to keep
packages; until the release in which it shipped goes EOL, which I think
is completely reasonable.