Revert "live: remove serial line from isolinux.cfg"
This reverts commit 0f1183e9ccca9279feff138576f516942e17b0a3.

Apparently removing the serial line means that we no longer get serial
console on a VM when doing something like `virt-install --graphics=none`.
Let's revert this change for now (i.e. go back to how it was before) and
make a subsequent change in the future that tries to handle all cases
better.