coreos-growpart: avoid creating ambivalent filesystems
If blkid finds a partition containing superblocks from multiple
filesystem types, it'll refuse to report anything about that partition.
If we growpart /sysroot into space that was previously used by ZFS, we'll
pick up a ZFS superblock and then fail to detect the filesystem type for
resizing.

Fix this by probing the filesystem type before running growpart.  After
running growpart, but before resizing, run wipefs(8) on the newly-grown
partition, asking it to delete all superblocks for filesystem types other
than the one we just detected.

This is kind of hacky.  We can't guarantee that wipefs will only touch
the unwritten space past the end of the existing filesystem.  Presumably
the existing filesystem contents are safe (either they come from an
image, or Ignition has just created the filesystem), but there's no
guarantee that wipefs won't detect an errant superblock signature and
thus modify the filesystem while it's mounted.  Of course, in that case
we'd already have an ambivalent filesystem and thus it wouldn't mount
properly in the future anyway.

Also exit early if the filesystem type is unrecognized.

Fixes coreos/fedora-coreos-tracker#500.