coreos-copy-firstboot-network.service: Update GPT generator unit name
The GPT setup unit is simplified to reflect the fact that this unit will
only be called once for the disk with the partition labeled 'boot'.

This keeps the old unit name for now for compatibility. This will be
safe to remove once the dracut-ignition changes from [1] are available
in the latest ignition RPM.

[1] https://github.com/coreos/ignition-dracut/pull/191