overrides: add fedora-coreos-pinger-0.0.4-3.fc32
Add the recent non-modular fedora-coreos-pinger build to
overides, removing our dependency on modular content. Part
of fixing https://github.com/coreos/fedora-coreos-tracker/issues/525.

Overriding is necessary as there is no Bodhi update yet.
The Bodhi update is pending a Fedora admin building with the
side tag to make the package available in F32. Permissions for
us to create the side tag are also pending, see:
https://pagure.io/releng/issue/9229. Once either of the two
pending above are completed, we'll be able to create and add
the Bodhi update here (see TODO in the manifest).