overrides: fast track microcode_ctl for CVE-2020-0543
Fast-track a build of microcode_ctl that is part of a set of fixes
for a recent CVE (CVE-2020-0543). The bug for the CVE is BZ1827165
which links to two Fedora specific bugs: BZ1845629 (kernel) and
BZ1845630 (microcode_ctl). The kernel update (kernel-5.6.18-300.fc32)
already made it to stable. The microcode_ctl update has not yet.

https://bodhi.fedoraproject.org/updates/FEDORA-2020-7bb2398b6b

Note that microcode_ctl only applies to x86_64 so we are explicitly
not updating the other architecture overrides files.