40ignition-ostree: relabel /root after copying skel files
Otherwise they can end up unlabeled. We don't hit this right now because
Ignition is just doing a blind relabeling on all of `/var/roothome`
whenever any user is configured (which is always because we have a base
config which configures the `core` user). We shouldn't rely on this
though:

https://github.com/coreos/ignition/pull/996

And more generally, each piece of software should be responsible for
relabeling the bits that it touched.