live-generator: Add Options=ro to our mounts
The default is a writable mount, and right now the
kernel emits a warning when the mount succeeds
because the underlying block device is read-only.

Which...actually turns out to *also* be the cause behind
the races we've been seeing mounting the ISO
https://github.com/coreos/fedora-coreos-tracker/issues/523

I believe what's happening is one of the bits of udev
is probing the ISO and doing a read-only mount at the same time we're
trying to mount it.

From the kernel perspective it's fine to have multiple concurrent
read-only mounts, but because ours was requested as writable,
if we happend to be trying to do our mount at the same time
as the probing the kernel will reject it.