40ignition-ostree: add coreos-inject-rootmap.service
This implements the rootmap functionality that figures out all the
dependencies required to find `/sysroot`, and injects them into the BLS
config. For more information, see:

https://github.com/coreos/fedora-coreos-tracker/issues/94#issuecomment-650433166

The `rdcore` code supports RAID and LUKS devices, though the latter
needs a new Clevis release with the following patches to be fully
supported:

https://github.com/latchset/clevis/pull/211
https://github.com/latchset/clevis/pull/217

This also implements the `root=UUID=$uuid` inject patch proposed in
https://github.com/coreos/fedora-coreos-tracker/issues/465.

On its own, this unlocks reprovisioning FCOS with root on a RAID device,
or e.g. in-place reprovisioning of root on btrfs.

Closes: https://github.com/coreos/fedora-coreos-tracker/issues/465
Closes: https://github.com/coreos/fedora-coreos-tracker/issues/94