live-generator: Avoid tmpfs/overlayfs, add stronger deps
This has many things rolled into one commit because I just
needed to test them together.  The high level goal is SELinux
enforcing for RHCOS and to mitigate a boot time race (still
not fully debugged, ended up adding `After=systemd-udev-settle.service`)

The SELinux enforcing one turned up a huge mess because as
it happens SELinux will nuke all labels on `tmpfs` after policy
is loaded.  Which is very problematic in general (FCOS included)
but FCOS was papering over this problem by using the systemd `relabel-extra.d`
bits.

But we don't really want to use that anymore IMO - we should
have a clean model where files are always labeled correctly in the initramfs
before we switch root.  Anything else going to lead to pain.

In order to work around the SELinux/`tmpfs` bug, instead make a loopback-mounted
`xfs` filesystem (on `tmpfs`).

When the kernel is fixed to retain labels in `tmpfs` we can drop that
hack.