live: mount ISO squashfs directly out of the rootfs image
cosa now uses identical initramfs and rootfs images in the PXE and ISO
artifacts.  Instead of expecting /root.squashfs, find the rootfs image
in the ISO pxeboot directory and mount the squashfs from a fixed offset
in the cpio archive.  Use the existence of the coreos.liveiso= karg,
rather than a flag file, to detect that we're ISO-booted.  In ISO mode,
use bsdtar to read the osmet files out of the rootfs image.